HARDWARE: remove build system requirements
This introduces png_hardware.c which is pretty much a copy of pngpriv.h
but then moves the hardware-specific parts to the arch/arch_init.c
files.

The change is extensive by line numbers but quite simple by effect;
since we have established that all the hardware-specific files can be
compiled on all platforms it is no longer necessary to have any hardware
specific tests in the main part of libpng.

This was easy for the filter code but the palette code requires work
within the main body of libpng, for this reason this is a
work-in-progress and therefore a draft pull request.  Some amount of
testing is important at this point as my cross-platform tests are
limited.  Nevertheless this can be included in a 1.8 alpha; it meets all
the requirements of an alpha, it just needs a little review from parties
who can do the cross platform stuff on platforms other than intel and
arm.

Signed-off-by: John Bowler <jbowler@acm.org>