Optimize build size when no target code
This factors the original pnghardware.c into a header and the actual
implementation.

The header is pngtarget.h.  It sets general purpose macros prefixed with
PNG_TARGET_ to set up general changes to pngstruct.h and the actual
code in png*.c.  None of this depends on the actual target; any target,
new or old, can implement the "optimizations".

The code is in pngsimd.c.  If target-specific support is available it
includes a target-specific body of code then provides the additional
functions (all general purpose prefixed by png_target_) to implement the
target-specific code by calling the functions from the included
"target-specific body of code".

This means that new implementations do not modify the main body of the
libpng code except for a one-line addition of the include of a
macro-only file in pngtarget.h.  If someone objects to the code for some
reason they can just comment out that line and not a single bit of it
will even get parsed by the compiler!

That, however, should never be necessary: the "check" header file just
defines minimal macros and pngtarget.h verifies that they are
consistent.  I believe it is almost impossible for one "check" header to
clobber another; the compiler will produce a diagnostic.

The approach also means that the checks get performed in a consistent
way that allows minimization of the overhead of target specific code
support on targets which have none.

Signed-off-by: John Bowler <jbowler@acm.org>