doc: Move all manuals to the manuals/ directory
Create the manuals/ directory for storing the software manuals and the
man pages. The upcoming Markdown versions of these document are also
planned to be added here.

Move in the manual files as follows:
 * Move libpng.3 and libpng.5 to manuals/libpng.3 and manuals/libpng.5
 * Move libpng-manual.txt to manuals/libpng-manual.txt
 * Move INSTALL.txt to manuals/libpng-install.txt
 * Remove libpngpf.3

Finally, update the README file, as well as the makefiles and the CMake
file, in order to reflect these changes.