ARM Ltd palette SIMD code
This enables the ARM Ltd SIMD code to assist "png_expand_palette".  The
code has been refactored so that while it does not change the disruption
to the main libpng code, in particular pngrtran.c, is minimized and
generalized.

It should now be possible for any manufacturer to optimize the same
transform (expansion of a color-mapped PNG to RGB or RGB8) without new
manufacturer favouring code in the main body of libpng.

Signed-off-by: John Bowler <jbowler@acm.org>