Use standard C attributes
ISO have recently standardized a way of using attributes.  This allows
removal of the existing compiler-specific hacks because the existing
supported compilers support the new standard.  This allows enables
market competition in the compiler world by removing support for
specific compilers in this case.

The changes are detailed in these commits:

commit 51b6189d3c21711bbd95538883c0aca3bb61dc67
commit 8942be90442e7a59260b01e49cdb4a3bc50c80b8