Full compile time support for target-specific code
This removes the need for build configuration of target-specific code
such as SIMD enhancements are machine specific coding (e.g. Intel's
crc32 instruction).

The extensible framework based on checks in the new pngtarget.h allows
per-target description of requirements and capabilities using
PNG_TARGET_ macros defined by the core code and per-target
implementation within, currently, a single core file pngsimd.c

This ensures that no special configuration in the build system is
required on any system which is compiling for the build host (where we
can reasonably assume that the compiler is correct set up) or for
cross-builds so long as the toolchain - the cross compiler and any
ancilliary programs it requires - are correct set up.

See pngtarget.h and pngsimd.c for more documentation.

Signed-off-by: John Bowler <jbowler@acm.org>