build system: require ISO-C when building libpng
Previously the build systems would accept any old compiler and those
compilers would define any old symbol or macro name.  ISO (well, ANSI)
defined a strict set of symbols/macros which may be defined by the
implementation leaving the remainder for application programs.

Adding a requirement for an ISO-C compiler (any version) ensures that a
check for name clashes on one compiler (with a sufficient level of
diagnostics) works for every ISO-C compiler.  In other words if it
builds on one it builds on all.

The check is in pngpriv.h; the same restrictions do not apply to code
that calls the public interface of libpng, only to the compiler used to
build ligpng.

Signed-off-by: John Bowler <jbowler@acm.org>