code cleanup: Remove 10700 checkes where possible.
This removes certain of the 10700 checks which can safely be removed as
they depend on the major require in libpng-1.7 which is not happening
in libpng-1.8.

The checks in pngvalid.c remain because it may (should) be possible to
fix or partially the underlying libpng bugs in 1.8.  The checks document
what those bugs are and ideally the 10700 branch of the #if's can be
used after a fix.

Signed-off-by: John Bowler <jbowler@acm.org>