chore: Clean up the version 10700 checks where possible
This removes some of the 10700 checks that can safely be removed as
they depend on the major require in libpng-1.7 which is not happening
in libpng-1.8.

The checks in pngvalid.c remain, because it may (should) be possible to
fix (or partially fix) the underlying libpng bugs in 1.8.  The checks
document what those bugs are.  Ideally, the 10700 branch of the #if's
can be used after a fix.

Reviewed-by: Cosmin Truta <ctruta@gmail.com>
Signed-off-by: John Bowler <jbowler@acm.org>
Signed-off-by: Cosmin Truta <ctruta@gmail.com>