QD fix: use png_aligncastconst for the upcast in NEON palette
This changes the arm/palette_neon_intrinsics.c code from using a bare
cast to using the portable macro png_aligncastconst.  This removes the
-Wcast-align warning.

The cast is safe because the pointer was allocated by malloc(1024) and
malloc is required to return a pointer with sufficient alignment for
anything the allocation can accomodate.

Nevertheless this is a quick and dirty fix; the use of png_bytep in the
png_struct declaration of the pointer is entirely machine specific,
other architectures may need different types and consequently the only
type which is portable to other companies CPUs is (void*).

Signed-off-by: John Bowler <jbowler@acm.org>