refactor: Deduplicate the interlace arrays `png_pass_*`
Move the definitions of interlace arrays to the top of their respective
translation units, to reduce their repeated copying in various places of
the codebase.

TODO:
These arrays still exist in three copies, in pngpread.c, in pngrutil.c,
and in pngwutil.c, and they should be further moved to a common place.