api(Windows): Remove the ordinals attached to exported functions
In order to make ABI compatibility guarantees about the Windows DLL,
we had to keep maintaining the API ordinals consistently. As we are
now designing a fresh new ABI, we use this opportunity to stop using
ordinals altogether.

Exporting API functions by ordinal numbers instead of names was a
necessity in the early (16-bit) versions of Windows, by design, to
minimize the memory requirements. Although using ordinals was still
possible in all subsequent versions of Windows, it was no longer
necessary from the 32-bit versions of Windows onwards.

The libpng DLL will still have ordinals, assigned arbitrarily, but
the dynamic linking of applications to this DLL will be done by name.
(The applications that use the zlib DLL operate in the same manner.)

Reviewed-by: John Bowler <jbowler@acm.org>
Signed-off-by: Cosmin Truta <ctruta@gmail.com>