testing: update pngtest FP and fixed API checks
pngtest.c is changed to check the floating point APIs by default as
these are likely to be the most used and in a number of cases they still
use the fixed point APIs.

contrib/conftest/float.dfa is added to turn off the fixed-point APIs and
these force use of the available floating point APIs in all test
programs when used as the build configuration.

Signed-off-by: John Bowler <jbowler@acm.org>