test: Improve tests/pngtest-all - work without WRITE
contrib/conftest/read-full.dfa failed on the bad palette tests in
tests/pngtest-all because the error message for writing a PNG with a
bad palette index was not output.

pngtest-all now checks for PNG_WRITE_SUPPORTED and, if not set, skips
the check for the write error message.

This is a cherry-pick of commit f276c3383f64978343c00df744cde9b9c12a86af
from branch 'develop'.

Reviewed-by: Cosmin Truta <ctruta@gmail.com>
Signed-off-by: John Bowler <jbowler@acm.org>
Signed-off-by: Cosmin Truta <ctruta@gmail.com>