api(apng): Handle APNG and Apple chunks correctly
pngset.c now includes the APNG chunks in the list of chunks to ignore
when `png_set_keep_unknown_chunks` is called to ignore all ancillary
chunks.

contrib/tools/pngcp.c uses this to correctly copy APNG chunks and other
known unsafe-to-copy chunks that pngcp can safely copy.  The change also
means that Apple's private banding chunk is not copied; it apparently
relies on the exact layout of the IDAT data.

Reviewed-by: Cosmin Truta <ctruta@gmail.com>
Signed-off-by: John Bowler <jbowler@acm.org>
Signed-off-by: Cosmin Truta <ctruta@gmail.com>