fix(apng): Correct `info_rowbytes` in `png_read_reinit`
In the case where the application has not called `png_read_update_info`,
`png_ptr->info_rowbytes` is 0 and should not be updated.

In the case where the application has called `png_read_update_info`,
the value of `info_ptr->rowbytes` needs to be updated as well as
`png_ptr->info_rowbytes`, so that `png_get_rowbytes` works correctly.

Co-authored-by: Maarten Bent <MaartenBent@users.noreply.github.com>
Reviewed-by: Cosmin Truta <ctruta@gmail.com>
Signed-off-by: John Bowler <jbowler@acm.org>
Signed-off-by: Cosmin Truta <ctruta@gmail.com>