Simplify png_sig_cmp
The png_sig_cmp function verifies that even if the passed amount of
bytes to check (num_to_check) is too large, no read overflow occurs.

While the first num_to_check verification is easy to read, the second
one is not.

Since another check verifies that start is not larger than 7, the
check can be rewritten to bring `num_to_check`, `8` and `start` in
same order as in the statement within the if-block. This makes it very
easy to see that it's pretty much a `MIN()` statement.

With this reordering, the initial num_to_check verification becomes
obsolete as well, allowing compilers to optimize much better. Both,
clang and gcc, reduce the amount of required jump instructions and the
library size shrinks by 16 (clang) or 64 (gcc) bytes on x86_64.