Require zlib 1.2.8 -- shed maintenance
There are some proposals to move to a new zlib backend.

My first step in writing a patch for this was to shed some of the
compatibility if statements.

Zlib 1.2.8 was release in 2013 according to the archive history
https://zlib.net/fossils/

I think that 12 years is pretty generous and should help shed some mental
load